Added files via upload
Hi Guy I finally managed to have the game to stop at the end of the round of the first player which reach 13 brains. Note in this game version line 158 i changed the value if (playerScore[currentPlayer] >= 4) // winner? rather than 13 as otherwise it will take to long to check value